feature: Create tests to validate user and film fields
