+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -6,7 +6,6 @@ import lombok.Setter;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
 import java.time.LocalDate;
 
 @Getter
@@ -16,7 +15,6 @@ public class User {
     private int id;
     @Email
     private String email;
-    @NotEmpty
     @NotBlank
     private final String login;
     private String name;
